+++ src/aoiro/reader/_io.py
